Improve satellites router
