+++ pkg/satellite/api/applications/router_test.go
@@ -1,0 +1,12 @@
+package applications
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestNewRouter(t *testing.T) {
+	router := Mount(Config{}, nil)
+	require.NotNil(t, router)
+}
