+++ pkg/satellite/api/clusters/clusters.go
@@ -1,4 +1,4 @@
-package router
+package clusters
 
 import (
 	"net/http"
