+++ pkg/satellite/api/teams/router.go
@@ -1,0 +1,27 @@
+package teams
+
+import (
+	"github.com/kobsio/kobs/pkg/kube/clusters"
+
+	"github.com/go-chi/chi/v5"
+)
+
+type Config struct{}
+
+type Router struct {
+	*chi.Mux
+	config         Config
+	clustersClient clusters.Client
+}
+
+func Mount(config Config, clustersClient clusters.Client) chi.Router {
+	router := Router{
+		chi.NewRouter(),
+		config,
+		clustersClient,
+	}
+
+	router.Get("/teams", router.getTeams)
+
+	return router
+}
