+++ pkg/satellite/api/teams/router_test.go
@@ -1,0 +1,12 @@
+package teams
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestNewRouter(t *testing.T) {
+	router := Mount(Config{}, nil)
+	require.NotNil(t, router)
+}
